+++ aiohttp_sqlalchemy/web_handlers.py
@@ -67,14 +67,14 @@ class OffsetPaginationMixin(ahth.PaginationMixin, SelectStatementMixin):
             self.context['items'] = page.items
 
             if page.next:
-                next_ = str(page.next)
-                self.context['next_url'] = route.url_for(page_key=next_)
+                kw = {'page_key': page.next}
+                self.context['next_url'] = route.url_for().with_query(kw)
             else:
                 self.context['next_url'] = page.next
 
             if page.previous:
-                previous_ = str(page.previous)
-                self.context['previous_url'] = route.url_for(page_key=previous_)
+                kw = {'page_key': page.previous}
+                self.context['previous_url'] = route.url_for().with_query(kw)
             else:
                 self.context['previous_url'] = page.previous
 
@@ -83,12 +83,12 @@ class PrimaryKeyMixin(ahth.PrimaryKeyMixin, SAModelMixin, metaclass=ABCMeta):
     sa_pk_attr: Any = getattr(SAModelMixin.sa_model, 'pk', None)
 
 
-class ItemAddMixin(SAModelMixin, ahth.ItemMixin, metaclass=ABCMeta):
+class UnitAddMixin(SAModelMixin, ahth.ItemMixin, metaclass=ABCMeta):
     def sa_add(self, *, key: Optional[str] = None) -> None:
         self.get_sa_session(key).add(self.item)
 
 
-class ItemDeleteMixin(
+class UnitDeleteMixin(
     DeleteStatementMixin,
     PrimaryKeyMixin,
     metaclass=ABCMeta,
@@ -99,7 +99,7 @@ class ItemDeleteMixin(
             where(self.sa_pk_attr == self.pk)
 
 
-class ItemEditMixin(
+class UnitEditMixin(
     ahth.ItemMixin,
     UpdateStatementMixin,
     PrimaryKeyMixin,
@@ -111,7 +111,7 @@ class ItemEditMixin(
             where(self.sa_pk_attr == self.pk)
 
 
-class ItemViewMixin(
+class UnitViewMixin(
     ahth.ItemMixin,
     SelectStatementMixin,
     PrimaryKeyMixin,